+++ package.json
@@ -10,6 +10,8 @@
     "@types/node": "^12.0.0",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
+    "@types/react-router": "^5.1.14",
+    "@types/react-router-dom": "^5.1.7",
     "@types/styled-components": "^5.1.9",
     "date-fns": "^2.21.1",
     "eslint-config-prettier": "^8.2.0",
@@ -23,6 +25,7 @@
     "react-icons": "^4.2.0",
     "react-markdown": "^6.0.1",
     "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
     "react-spinners": "^0.10.6",
     "react-syntax-highlighter": "^15.4.3",
