+++ src/components/Todos.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Todos = () => {
+  return (
+    <div>
+      <h1>TODOS</h1>
+    </div>
+  )
+}
+
+export default Todos
